chore: rename packages back to single-spa-angular
Renaming everything back to single-spa-angular because the GitHub support
team has transferred permissions for single-spa-angular to @arturovt.
