+++ README.md
@@ -1,8 +1,8 @@
 <p align="center">
-  <img src="https://raw.githubusercontent.com/single-spa/single-spa-angular/refs/heads/v19.0.0/docs/assets/logo-256.png" alt="@single-spa-community/angular logo" width="256" height="256" />
+  <img src="https://raw.githubusercontent.com/single-spa/single-spa-angular/refs/heads/v19.0.0/docs/assets/logo-256.png" alt="single-spa-angular logo" width="256" height="256" />
 </p>
 
-# @single-spa-community/angular
+# single-spa-angular
 
 Helpers for building [single-spa](https://github.com/single-spa/single-spa) applications which use Angular.
 
@@ -22,10 +22,10 @@ The package has been republished under the `@single-spa-community` scope due to 
 
 ## Compatibility
 
-| `@single-spa-community/angular` | Angular |
-| ------------------------------- | ------- |
-| `21.x`                          | `21.x`  |
-| `20.x`                          | `20.x`  |
-| `19.x`                          | `19.x`  |
+| `single-spa-angular` | Angular |
+| -------------------- | ------- |
+| `21.x`               | `21.x`  |
+| `20.x`               | `20.x`  |
+| `19.x`               | `19.x`  |
 
 > **Note:** This package follows a lock-step versioning strategy — the major version is aligned with Angular's major version to make compatibility obvious at a glance. This means a major bump does not necessarily indicate a breaking change in this package's own API.
