+++ apps/navbar/src/app/app.config.ts
@@ -1,7 +1,7 @@
 import { APP_BASE_HREF } from '@angular/common';
 import { ApplicationConfig } from '@angular/core';
 import { provideRouter } from '@angular/router';
-import { provideSingleSpa } from '@single-spa-community/angular';
+import { provideSingleSpa } from 'single-spa-angular';
 
 import { routes } from './app.routes';
