+++ apps/root-config/src/index.ejs
@@ -53,10 +53,10 @@
 
           "@angular/elements": "http://localhost:4600/es2022/angular-elements.js",
 
-          "@single-spa-community/angular/internals": "http://localhost:4600/es2022/single-spa-community-angular-internals.js",
-          "@single-spa-community/angular": "http://localhost:4600/es2022/single-spa-community-angular.js",
-          "@single-spa-community/angular/elements": "http://localhost:4600/es2022/single-spa-community-angular-elements.js",
-          "@single-spa-community/angular/parcel": "http://localhost:4600/es2022/single-spa-community-angular-parcel.js"
+          "single-spa-angular/internals": "http://localhost:4600/es2022/single-spa-angular-internals.js",
+          "single-spa-angular": "http://localhost:4600/es2022/single-spa-angular.js",
+          "single-spa-angular/elements": "http://localhost:4600/es2022/single-spa-angular-elements.js",
+          "single-spa-angular/parcel": "http://localhost:4600/es2022/single-spa-angular-parcel.js"
         }
       }
     </script>
@@ -87,10 +87,10 @@
 
           "@angular/elements": "http://localhost:4600/es2022/angular-elements.min.js",
 
-          "@single-spa-community/angular/internals": "http://localhost:4600/es2022/single-spa-community-angular-internals.min.js",
-          "@single-spa-community/angular": "http://localhost:4600/es2022/single-spa-community-angular.min.js",
-          "@single-spa-community/angular/elements": "http://localhost:4600/es2022/single-spa-community-angular-elements.min.js",
-          "@single-spa-community/angular/parcel": "http://localhost:4600/es2022/single-spa-community-angular-parcel.min.js"
+          "single-spa-angular/internals": "http://localhost:4600/es2022/single-spa-angular-internals.min.js",
+          "single-spa-angular": "http://localhost:4600/es2022/single-spa-angular.min.js",
+          "single-spa-angular/elements": "http://localhost:4600/es2022/single-spa-angular-elements.min.js",
+          "single-spa-angular/parcel": "http://localhost:4600/es2022/single-spa-angular-parcel.min.js"
         }
       }
     </script>
