+++ libs/single-spa-angular/elements/ng-package.json
@@ -2,6 +2,6 @@
   "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
   "lib": {
     "entryFile": "src/index.ts",
-    "flatModuleFile": "single-spa-community-angular-internals"
+    "flatModuleFile": "single-spa-angular-elements"
   }
 }
