+++ libs/single-spa-angular/elements/src/types.ts
@@ -1,6 +1,6 @@
 import type { ApplicationRef, NgModuleRef } from '@angular/core';
 import type { NgElement } from '@angular/elements';
-import type { BaseSingleSpaAngularOptions } from '@single-spa-community/angular/internals';
+import type { BaseSingleSpaAngularOptions } from 'single-spa-angular/internals';
 
 export interface BootstrappedSingleSpaAngularElementsOptions extends BaseSingleSpaAngularOptions {
   ngModuleRefOrAppRef: NgModuleRef<any> | ApplicationRef | null;
