+++ libs/single-spa-angular/index.ts
