+++ libs/single-spa-angular/internals/src/dom.ts
