+++ libs/single-spa-angular/internals/src/index.ts
