+++ libs/single-spa-angular/internals/src/types.ts
