+++ libs/single-spa-angular/ng-package.json
@@ -2,7 +2,7 @@
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
   "dest": "../../lib",
   "lib": {
-    "flatModuleFile": "single-spa-community-angular",
+    "flatModuleFile": "single-spa-angular",
     "entryFile": "index.ts"
   }
 }
