+++ libs/single-spa-angular/package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../node_modules/ng-packagr/ng-package.schema.json",
-  "name": "@single-spa-community/angular",
+  "name": "single-spa-angular",
   "version": "21.0.1",
   "description": "Helpers for building single-spa applications which use Angular 2",
   "schematics": "./schematics/schematics.json",
