+++ libs/single-spa-angular/parcel/src/index.ts
