+++ libs/single-spa-angular/parcel/src/parcel.component.ts
