+++ libs/single-spa-angular/parcel/src/parcel.module.ts
