+++ libs/single-spa-angular/project.json
@@ -1,0 +1,17 @@
+{
+  "name": "single-spa-angular",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "sourceRoot": "libs/single-spa-angular/src",
+  "prefix": "single-spa-angular",
+  "targets": {
+    "build": {
+      "executor": "@nx/angular:package",
+      "options": {
+        "tsConfig": "libs/single-spa-angular/tsconfig.lib.json",
+        "project": "libs/single-spa-angular/ng-package.json"
+      }
+    }
+  },
+  "tags": []
+}
