+++ libs/single-spa-angular/src/platform-providers.ts
