+++ libs/single-spa-angular/src/providers.ts
