+++ libs/single-spa-angular/src/public_api.ts
