+++ libs/single-spa-angular/src/run-outside-angular.ts
