+++ libs/single-spa-angular/src/single-spa-angular.ts
@@ -1,6 +1,6 @@
 import type { ApplicationRef, NgModuleRef, NgZone } from '@angular/core';
 import type { LifeCycles } from 'single-spa';
-import { getContainerElementAndSetTemplate } from '@single-spa-community/angular/internals';
+import { getContainerElementAndSetTemplate } from 'single-spa-angular/internals';
 
 import { SingleSpaPlatformLocation } from './platform-providers';
 import type { SingleSpaAngularOptions, BootstrappedSingleSpaAngularOptions } from './types';
