+++ libs/single-spa-angular/src/types.ts
@@ -1,6 +1,6 @@
 import type { NgModuleRef, Type, NgZone, ApplicationRef } from '@angular/core';
 import type { AppProps } from 'single-spa';
-import type { BaseSingleSpaAngularOptions } from '@single-spa-community/angular/internals';
+import type { BaseSingleSpaAngularOptions } from 'single-spa-angular/internals';
 
 export interface SingleSpaAngularOptions<
   T = Record<string, unknown>,
