+++ libs/single-spa-angular/tsconfig.json
