+++ libs/single-spa-angular/tsconfig.lib.json
