+++ libs/single-spa-angular/webpack/index.spec.ts
