+++ libs/single-spa-angular/webpack/index.ts
