+++ libs/single-spa-angular/webpack/webpack-5/remove-mini-css-extract.ts
