+++ schematics/ng-add/_files/extra-webpack.config.js.template
@@ -1,4 +1,4 @@
-const singleSpaAngularWebpack = require('@single-spa-community/angular/lib/webpack').default;
+const singleSpaAngularWebpack = require('single-spa-angular/lib/webpack').default;
 
 module.exports = (config, options) => {
   const singleSpaWebpackConfig = singleSpaAngularWebpack(config, options);
